requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase>=2.3.4
python-dotenv==1.0.0

# Document Processing
PyMuPDF==1.23.26
python-multipart==0.0.6

# AI/ML Dependencies
openai>=1.3.0
google-generativeai>=0.3.2
sentence-transformers>=2.2.2
numpy>=1.24.0
requests>=2.31.0

# Additional utilities
aiofiles==23.2.0
pydantic==2.5.0
httpx==0.25.2
